% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_corpus.R
\name{error_corpus}
\alias{error_corpus}
\title{Check and Handle Errors Related to the Corpus}
\usage{
error_corpus(corpus, from, to, resolution)
}
\arguments{
\item{corpus}{A character string. The corpus specified by the user.}

\item{from}{An integer. Starting year.}

\item{to}{An integer. End year.
The finest available resolution for the corpus selected can be found
in the \code{resolution} column of the \code{list_corpora} dataset.}

\item{resolution}{A character string. Can only be "daily", "monthly" or
"yearly".
The finest available resolution for the corpus selected can be found
in the \code{resolution} column of the \code{list_corpora} dataset.}
}
\value{
This function does not return a value but may issue warnings or
errors if any issues are found.
}
\description{
This internal function checks for errors in the specified corpus,
including validity of the corpus required,
reliability of the specified time range,
and compatibility of the requested resolution
with the corpus's available resolutions.
}
\details{
This function performs the following checks:
\itemize{
  \item{Validity of Corpus Code:}{
   Ensures that the specified \code{corpus} is a valid corpus.
  }
  \item{Reliability of Time Range:}{
   Checks whether the specified time range (\code{from} to \code{to})
   falls within the reliable range of the specified corpus.
  }
  \item{Compatibility of Resolution:}{
   Verifies that the requested resolution (\code{resolution})
   is compatible with the available resolutions of the specified corpus.
 }
}
}
\examples{
\dontrun{
  error_corpus("lemonde", 1940, 2025, "monthly")
}

}
\keyword{error}
\keyword{internal,}
