% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_param.R
\name{error_param}
\alias{error_param}
\title{Internal function to check for errors in the parameters}
\usage{
error_param(info_corpus, keyword, corpus, from, to, resolution, n_of_level)
}
\arguments{
\item{info_corpus}{A data frame containing information about the selected
corpus.}

\item{keyword}{A character string specifying the keyword for analysis.}

\item{corpus}{A character string specifying the code of the corpus.}

\item{from}{An integer or "earliest" representing the starting year
for the analysis.}

\item{to}{An integer or "latest" representing the ending year
for the analysis.}

\item{resolution}{A character string specifying the desired resolution for
the analysis (e.g., "daily", "monthly", "yearly").}

\item{n_of_level}{A character string specifying the type of object to
compute the number of occurrences for or the level at which to compute
co-occurrences.}
}
\value{
This function does not return a value but may issue warnings or
errors if any issues are found.
}
\description{
This internal function checks for errors in the specified parameters,
including the validity of the time range, the keyword, and the resolution,
based on the provided information about the corpus.
}
\details{
This function performs the following checks:
 \itemize{
  \item{Validity of Time Range:}{Ensures that the specified time range
  (\code{from} to \code{to}) is numeric and valid.}
  \item{Validity of Keyword:}{Checks whether the specified \code{keyword} is
  a character string and not a character vector.}
  \item{Reliability of Corpus:}{Verifies that the specified time range falls
  within the reliable range of the given corpus.}
  \item{Validity of Resolution:}{Ensures that the specified
  \code{resolution} is one of "daily", "monthly", or "yearly" and is
  compatible with the corpus's available resolutions.}
}
}
\keyword{error}
\keyword{internal,}
