% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gallicagram_all.R
\name{gallicagram_all}
\alias{gallicagram_all}
\title{Occurrences of a keyword in all Gallicagram corpora}
\usage{
gallicagram_all(keyword)
}
\arguments{
\item{keyword}{A character string. Keyword to search.
The string cannot contain more words than the \code{max_length}
for this corpus, as indicated in the \code{list_corpora} dataset.}
}
\value{
A tibble. With the \code{keyword},
the number of occurrences (\code{n_occur}) or co-occurrences
(\code{n_cooccur}), the total number of ngrams or aritcles
over the period (\code{n_total}),
the proportion of occurrences or co-occurrences of the keyword(s) over the
period of a given observation (\code{prop_occcur} or \code{prop_coocccur}),
either info about whether the total number is a number of grams or articles
(\code{n_of}) or about
the syntagma at which the co-occurrences are computed (\code{cooccur_level}),
the date at the beginning of the period of a given observation (\code{date}),
the \code{corpus}, the \code{resolution},
the \code{year} and
potentially the \code{month} and \code{day} of the observation.
}
\description{
Retrieves the yearly proportion of occurrences of a keyword in the lemonde,
books and press corpora, for the longest time range possible
while data remaining reliable.
}
\examples{
\dontrun{
  gallicagram_all("président")
}
}
