% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gallicagram_follows.R
\name{gallicagram_follows}
\alias{gallicagram_follows}
\title{Find the word the most often associated with a n-gram}
\usage{
gallicagram_follows(
  keyword,
  corpus = "lemonde",
  from = 1945,
  to = 2022,
  nb_joker = 20,
  after = TRUE
)
}
\arguments{
\item{keyword}{A character string. Keyword to search.}

\item{corpus}{A character string. The corpus to search. Takes the following
values: "press" for historical press, "books" for Gallica books,
"lemonde" for Le Monde newspaper articles.}

\item{from}{A number. Starting year.}

\item{to}{A number. End year.}

\item{nb_joker}{An integer. The number of most frequent occurrences to
return.}

\item{after}{A boolean. Whether to consider only words following the keyword
and not those preceding. Set to \code{TRUE} by default.}
}
\value{
A tibble. With the most frequent ngrams containing the keyword
mentioned (\code{ngram}) and the number of occurrences over the period
(\code{nb_occur}).
}
\description{
Returns the word the most frequently following a given ngram over the period.
It can also include the words that precede by setting
\code{after = FALSE}.

It corresponds to the 'Joker' function on the Gallicagram app
and is analogous to the 'Joker' function on Ngram Viewer.

For instance "camarade" is often followed by "staline" or "khrouchtchev" in
Le Monde. The function returns the most frequent ngrams of the form
"camarade *". Setting \code{after = FALSE} also includes the most frequent
ngrams of the form "* camarade".

The keyword can be a 2-gram in the "books" and "press" corpora and a 3-gram
in the "lemonde" corpus. Searching the "press" corpus can require a long
running time.
}
\examples{
gallicagram_follows("camarade")
}
