% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gallicagram_lexicon.R
\name{gallicagram_lexicon}
\alias{gallicagram_lexicon}
\title{Occurrences of a list of keywords in a Gallicagram corpus}
\usage{
gallicagram_lexicon(
  lexicon,
  corpus = "lemonde",
  from = 1945,
  to = 2022,
  resolution = "monthly"
)
}
\arguments{
\item{lexicon}{A character vector. Keywords to search. Can contain up to
a 3-gram in the "books" and "press" corpora and
a 4-gram in the "lemonde" corpus.}

\item{corpus}{A character string. The corpus to search. The list of
available corpora can be found in the \code{list_corpora} dataset.}

\item{from}{An integer. Starting year.}

\item{to}{An integer. End year.
The finest available resolution for the corpus selected can be found
in the \code{resolution} column of the \code{list_corpora} dataset.}

\item{resolution}{A character string. Can only be "daily", "monthly" or
"yearly".
The finest available resolution for the corpus selected can be found
in the \code{resolution} column of the \code{list_corpora} dataset.}
}
\value{
A tibble. With the first \code{keyword} in the vector,
typically the one, the entire \code{lexicon},
the number of occurrences (\code{n_occur}) or co-occurrences
(\code{n_cooccur}), the total number of ngrams over the period of a
given observation (\code{n_grams} or \code{n_ngrams}),
the proportion of occurrences or co-occurrences of the keyword(s) over the
period of a given observation (\code{prop_occcur} or \code{prop_coocccur}),
the date at the beginning of the period of a given observation (\code{date}),
the \code{corpus}, the \code{resolution},
the \code{year} and
potentially the \code{month} and \code{day} of the observation.
}
\description{
Retrieves the proportion of occurrences of a list of keywords
in one of the corpora (historical press, Gallica books, Le Monde newspaper)
by year, month or day.
}
\details{
This function sums the outputs of calls of \code{gallicagram} obtained for
each keyword in the vector.

Can typically be used with the function \code{get_same_stem()}.
}
\examples{
\dontrun{
  gallicagram_lexicon(c("président", "présidentiel"))
}
}
