% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gallicagram_with.R
\name{gallicagram_with}
\alias{gallicagram_with}
\title{ngrams that most frequently contain a keyword in a Gallicagram corpus}
\usage{
gallicagram_with(
  keyword,
  corpus = "lemonde",
  from = 1945,
  to = 2022,
  n_results = 20,
  after = FALSE,
  length = 2
)
}
\arguments{
\item{keyword}{A character string. Keyword to search. Can be up to a 3-gram
in the "books" and "press" corpora and a 4-gram in the "lemonde" corpus.}

\item{corpus}{A character string. The corpus to search. Takes the following
values: "press" for historical press, "books" for Gallica books,
"lemonde" for Le Monde newspaper articles.}

\item{from}{An integer. Starting year.}

\item{to}{An integer. End year.}

\item{n_results}{An integer. The number of most frequently
associated words to return. \code{n_results} can also be set to "all" to
return all the available results.}

\item{after}{A boolean. Whether to consider only words following the keyword
and not those preceding. Set to \code{FALSE} by default.}

\item{length}{An integer. The length of the ngrams considered.
Can be up to 3 in the "books" and "press" corpora and 4 in the
"lemonde" corpus.}
}
\value{
A tibble. With the \code{n_results} most frequent ngrams containing
the \code{keyword} searched (\code{ngram})
and the number of occurrences over the period (\code{n_occur}).
It also returns the input parameters
\code{keyword}, \code{corpus}, \code{from} and \code{to}.
}
\description{
Returns the most frequent ngrams containing a keyword over a given period.
}
\details{
This function corresponds to the \code{Joker} route of the API,
accessed through the 'Joker' function on the Gallicagram app.
When length = 1, it is analogous to the 'Joker' function on Ngram Viewer.

For instance "camarade" is often followed by "staline" or "khrouchtchev" in
Le Monde. The function returns the most frequent ngrams of the form
"camarade *" when setting \code{after = TRUE}. \code{after = FALSE} also
includes the most frequent ngrams of the form "* camarade".

Searching the "press" corpus can require a long running time.
}
\examples{
gallicagram_with("camarade", length = 2)
}
