% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gallicagraph.R
\name{gallicagraph}
\alias{gallicagraph}
\title{Automatically produces the relevant graph}
\usage{
gallicagraph(data, color = NULL)
}
\arguments{
\item{data}{A data frame produced by one of the \code{gallicagram} or
\code{gallicagram_coocur} functions or their lexicon counterparts
(or several of such data frames bound by rows).}

\item{color}{A variable to set colors for the graph}
}
\value{
A graph describing the evolution of the proportion of
(co-)occurrences of one or several keywords in one or several corpora.
}
\description{
Takes as input data produced by one of the \code{gallicagram} or
\code{gallicagram_coocur} functions or their lexicon counterparts and
produces the corresponding base graph.
}
\details{
This function can also be combined with faceting by adding for instance
\code{+ facet_wrap(~ keyword)} after calling the function.
}
\examples{
  gallicagram("président") |>
    gallicagraph()

  gallicagram_cooccur("président", "ancien") |>
    gallicagraph()
}
