% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gallicagraph_cooccur.R
\name{gallicagraph_cooccur}
\alias{gallicagraph_cooccur}
\title{Graph describing the evolution of the proportion of co-occurrences}
\usage{
gallicagraph_cooccur(data, color = NULL)
}
\arguments{
\item{data}{A data frame produced by the \code{gallicagram_cooccur} or
\code{gallicagram_cooccur_lexicon} functions
(or several of such data frames bound by rows).}

\item{color}{A variable to set colors for the graph}
}
\value{
A graph describing the evolution of the proportion of co-occurrences
of two keywords in one or several corpora.
}
\description{
Takes as input data produced by \code{gallicagram_cooccur} or
\code{gallicagram_cooccur_lexicon} and produces a graph
describing the evolution of the proportion of co-occurrences in this data set.
}
\details{
This function can also be combined with faceting by adding for instance
\code{+ facet_wrap(~ keyword)} after calling the function.
}
\examples{
  gallicagram_cooccur("président", "ancien") |>
    gallicagraph_cooccur()

  gallicagram_cooccur("président", "ancien") |>
    rbind(gallicagram_cooccur("président", "république")) |>
    gallicagraph_cooccur(color = keyword)

  gallicagram_cooccur("président", "ancien") |>
    rbind(gallicagram_cooccur("président", "république")) |>
    gallicagraph_cooccur() +
    ggplot2::facet_wrap(~ keyword)
}
