% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gallicagraph_occur.R
\name{gallicagraph_occur}
\alias{gallicagraph_occur}
\title{Graph describing the evolution of the proportion of occurrences}
\usage{
gallicagraph_occur(data, color = NULL)
}
\arguments{
\item{data}{A data frame produced by the \code{gallicagram} or
\code{gallicagram_lexicon} functions
(or several of such data frames bound by rows).}

\item{color}{A variable to set colors for the graph}
}
\value{
A graph describing the evolution of the proportion of occurrences
of one or several keywords in one or several corpora.
}
\description{
Takes as input data produced by \code{gallicagram} or
\code{gallicagram_lexicon} and produces a graph
describing the evolution of the proportion of occurrences in this data set.
}
\details{
This function can also be combined with faceting by adding for instance
\code{+ facet_wrap(~ keyword)} after calling the function.
}
\examples{
  gallicagram("président") |>
    gallicagraph_occur()

  gallicagram("président") |>
    rbind(gallicagram("république")) |>
    gallicagraph_occur(color = keyword)

  gallicagram("président") |>
    rbind(gallicagram("république")) |>
    gallicagraph_occur() +
    ggplot2::facet_wrap(~ keyword)

  gallicagram("président") |>
    rbind(gallicagram("président", corpus = "huma")) |>
    gallicagraph_occur(color = corpus)
}
