% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_same_stem.R
\name{get_same_stem}
\alias{get_same_stem}
\title{Finds words with the same stem as a keyword}
\usage{
get_same_stem(keyword)
}
\arguments{
\item{keyword}{A character string.}
}
\value{
A character vector. Words with the same stem has the \code{keyword}
}
\description{
Searches such words in the \code{rallicagram::word_lemma_stem} dataframe.
}
\examples{
\dontrun{
  get_same_stem("président")
}
}
