% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{list_corpora}
\alias{list_corpora}
\title{A data frame containing information about the available corpora}
\format{
A data frame with 17 rows and 7 variables:
\describe{
  \item{title}{The title of the corpus.}
  \item{start_reliable}{The year after which the corpus starts being reliable.}
  \item{end_reliable}{The year after which the corpus stops being reliable.}
  \item{nb_words}{The number of words in the corpus.}
  \item{code_api}{The code for to call the corpus in the API.}
  \item{max_length}{The maximum length of ngrams available.}
  \item{resolution}{The finest available resolution (e.g., daily, monthly, yearly).}
}
}
\usage{
data("list_corpora")
}
\description{
It includes the years for which the data is reliable,
the number of words in each corpora, the code to use in the API,
the maximum length of the ngrams, and the resolution.
}
\examples{
\dontrun{
  # Load the dataset
  data("list_corpora")
}

}
\keyword{datasets}
