% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_param.R
\name{prepare_param}
\alias{prepare_param}
\title{Internal function to prepare the parameters to use them in the API call.}
\usage{
prepare_param(keyword, corpus, from, to, resolution, n_of = "grams")
}
\arguments{
\item{keyword}{A character string. Keyword to search.
The largest number of words the keyword can contain can be found
in the \code{resolution} column of the \code{list_corpora} dataset.}

\item{corpus}{A character string. The corpus to search. The list of
available corpora can be found in the \code{list_corpora} dataset.}

\item{from}{An integer. Starting year.}

\item{to}{An integer. End year.
The finest available resolution for the corpus selected can be found
in the \code{resolution} column of the \code{list_corpora} dataset.}

\item{resolution}{A character string. Can only be "daily", "monthly" or
"yearly".
The finest available resolution for the corpus selected can be found
in the \code{resolution} column of the \code{list_corpora} dataset.}
}
\value{
A list.
Contains cleaned function parameters (`keyword`, `corpus` and `resolution`).
}
\description{
Checks potential errors, translates the parameters, set parameters for
the API call.
}
