% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_gallicagram.R
\name{read_data_gallicagram}
\alias{read_data_gallicagram}
\title{Internal function to read the output of the API call.}
\usage{
read_data_gallicagram(url)
}
\arguments{
\item{data}{The data produced by the Gallicagram API}
}
\value{
A tibble containing the raw data from the Gallicagram API.
}
\description{
Turns the output into a tibble
Only necessary to allow a longer timeout period.
}
