% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stopwords_gallica}
\alias{stopwords_gallica}
\title{A vector containing the most frequent words in the Gallica books corpus}
\format{
A length 1000 vector.
}
\usage{
stopwords_gallica
}
\description{
Sorted from most to less frequent word
}
\keyword{datasets}
