% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_gallicagram.R
\name{tidy_gallicagram}
\alias{tidy_gallicagram}
\title{Internal function to prepare the parameters to use them in the API call.}
\usage{
tidy_gallicagram(data, corpus, resolution)
}
\arguments{
\item{data}{A data frame produced by the Gallicagram API}

\item{corpus}{A character string. The corpus to search. Takes the following
values: "press" for historical press, "books" for Gallica books,
"lemonde" for Le Monde newspaper articles.}

\item{resolution}{A character string.
For Le Monde can be either "yearly", "monthly" or "daily".
For historical press can be either "yearly", "monthly".
For books can only only be. "yearly".}
}
\value{
A tibble. With the \code{keyword},
the number of occurrences (\code{nb_occur}) or co-occurrences
(\code{nb_cooccur}), the total number of ngrams over the period of a
given observation (\code{nb_grams} or \code{nb_ngrams}),
the proportion of occurrences or co-occurrences of the keyword(s) over the
period of a given observation (\code{prop_occcur} or \code{prop_coocccur}),
the date at the beginning of the period of a given observation (\code{date}),
the \code{source}, the \code{resolution},
the \code{year} and
potentially the \code{month} and \code{day} of the observation.
}
\description{
Internal function to prepare the parameters to use them in the API call.
}
