% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{word_lemma_stem}
\alias{word_lemma_stem}
\title{A data frame containing words, their lemmas and stem of the lemmas}
\format{
A data frame with 3 columns and 142,694 rows.
}
\usage{
word_lemma_stem
}
\description{
The lemmas come from the \href{http://www.lexique.org/}{Lexique} data base.
The stems are obtained by stemming these lemmas using the \code{SnowballC}
package.
}
\details{
This data frame can be used to get a vector of words that are closely related
to a keyword of interest.
}
\keyword{datasets}
