% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gallicagram.R
\name{gallicagram}
\alias{gallicagram}
\title{Retrieves the number of occurrences of a keyword in one of the corpora
(historical press, Gallica books, Le Monde newspaper)}
\usage{
gallicagram(
  keyword,
  corpus = "press",
  begin = 1789,
  end = 1950,
  resolution = "monthly"
)
}
\arguments{
\item{keyword}{A character string. Keyword to search.}

\item{corpus}{A character string. The corpus to search. Takes the following
values: "press" for historical press, "books" for Gallica books,
"lemonde" for Le Monde newspaper articles.}

\item{begin}{A number. Starting year.}

\item{end}{A number. End year.}

\item{resolution}{A character string.
For Le Monde can be either "yearly", "monthly" or "daily".
For historical press can be either "yearly", "monthly".
For books can only only be. "yearly".}
}
\value{
A tibble. With the `keyword`, the number of occurrences (`n`),
the total number of ngrams for the period for a given observation (`total`),
the date at the beginning of the period for a given observation (`date`),
the `source`, the `resolution`,
the `year` and potentially the `month` and `day` of the observation.
}
\description{
Retrieves the number of occurrences of a keyword in one of the corpora
(historical press, Gallica books, Le Monde newspaper)
}
\examples{
gallicagram("président")
}
