% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gallicagram.R
\name{gallicagram}
\alias{gallicagram}
\title{Occurrences of a keyword in a Gallicagram copus}
\usage{
gallicagram(
  keyword,
  corpus = "lemonde",
  from = "earliest",
  to = "latest",
  resolution = "monthly",
  n_of = "grams"
)
}
\arguments{
\item{keyword}{A character string. Keyword to search.
The largest number of words the keyword can contain can be found
in the \code{resolution} column of the \code{list_corpora} dataset.}

\item{corpus}{A character string. The corpus to search. The list of
available corpora can be found in the \code{list_corpora} dataset.}

\item{from}{An integer or "earliest". The starting year.
If set to "earliest", use the earliest date at which the data is reliable
for this corpus, as described in \code{list_corpora}.}

\item{to}{An integer or "latest". The end year.
If set to "latest", use the latest date at which the data is reliable
for this corpus, as described in \code{list_corpora}.}

\item{resolution}{A character string. Can only be "daily", "monthly" or
"yearly".
The finest available resolution for the corpus selected can be found
in the \code{resolution} column of the \code{list_corpora} dataset.}

\item{n_of}{A character string. The type of object to compute number of
occurrences for. If set to "grams", will compute the number of "grams" that
correspond to the keyword for the given period. If set to "article" (only
available for lemonde), will compute the number of articles that contain the
keyword for the given period.}
}
\value{
A tibble. With the \code{keyword},
the number of occurrences (\code{n_occur}) or co-occurrences
(\code{n_cooccur}), the total number of ngrams or aritcles
over the period (\code{n_total}),
the proportion of occurrences or co-occurrences of the keyword(s) over the
period of a given observation (\code{prop_occcur} or \code{prop_coocccur}),
either info about whether the total number is a number of grams or articles
(\code{n_of}) or about
the syntagma at which the co-occurrences are computed (\code{cooccur_level}),
the date at the beginning of the period of a given observation (\code{date}),
the \code{corpus}, the \code{resolution},
the \code{year} and
potentially the \code{month} and \code{day} of the observation.
}
\description{
Retrieves the proportion of occurrences of a keyword in one of the corpora
by year, month or day.
}
\details{
This function corresponds to the \code{Query} route of the API.

Information regarding available characteristics of the corpus can be found
in the \code{list_corpora} dataset.
}
\examples{
  gallicagram("président")

}
