% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gallicagram.R, R/gallicagram_cooccurrence.R
\name{gallicagram}
\alias{gallicagram}
\title{Retrieves the number of occurrences of a keyword in one of the corpora
(historical press, Gallica books, Le Monde newspaper) by year, month or day.}
\usage{
gallicagram(
  keyword,
  corpus = "press",
  from = 1789,
  to = 1950,
  resolution = "monthly"
)

gallicagram(
  keyword,
  corpus = "press",
  from = 1789,
  to = 1950,
  resolution = "monthly"
)
}
\arguments{
\item{keyword}{A character string. Keyword to search.}

\item{corpus}{A character string. The corpus to search. Takes the following
values: "press" for historical press, "books" for Gallica books,
"lemonde" for Le Monde newspaper articles.}

\item{from}{A number. Starting year.}

\item{to}{A number. End year.}

\item{resolution}{A character string.
For Le Monde can be either "yearly", "monthly" or "daily".
For historical press can be either "yearly", "monthly".
For books can only only be. "yearly".}
}
\value{
A tibble. With the `keyword`, the number of occurrences (`n`),
the total number of ngrams over the period of a given observation (`total`),
the proportion of occurrences of the keyword over the period of
a given observation (`prop`),
the date at the beginning of the period of a given observation (`date`),
the `source`, the `resolution`,
the `year` and potentially the `month` and `day` of the observation.

A tibble. With the `keyword`, the number of occurrences (`n`),
the total number of ngrams over the period of a given observation (`total`),
the proportion of occurrences of the keyword over the period of
a given observation (`prop`),
the date at the beginning of the period of a given observation (`date`),
the `source`, the `resolution`,
the `year` and potentially the `month` and `day` of the observation.
}
\description{
Retrieves the number of occurrences of a keyword in one of the corpora
(historical press, Gallica books, Le Monde newspaper) by year, month or day.

Retrieves the number of occurrences of a keyword in one of the corpora
(historical press, Gallica books, Le Monde newspaper) by year, month or day.
}
\examples{
gallicagram("président")
gallicagram("président")
}
