% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gallicagram_all.R
\name{gallicagram_all}
\alias{gallicagram_all}
\title{Occurrences of a keyword in all Gallicagram corpora}
\usage{
gallicagram_all(keyword)
}
\arguments{
\item{keyword}{A character string. Keyword to search.}
}
\value{
A tibble. With the \code{keyword},
the number of occurrences (\code{nb_occur}) or co-occurrences
(\code{nb_cooccur}), the total number of ngrams over the period of a
given observation (\code{nb_grams} or \code{nb_ngrams}),
the proportion of occurrences or co-occurrences of the keyword(s) over the
period of a given observation (\code{prop_occcur} or \code{prop_coocccur}),
the date at the beginning of the period of a given observation (\code{date}),
the \code{source}, the \code{resolution},
the \code{year} and
potentially the \code{month} and \code{day} of the observation.
}
\description{
Retrieves the yearly proportion of occurrences of a keyword in each of the
corpora for the longest time range for which the data is available
and reliable.
}
\examples{
gallicagram_all("président")
}
