% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gallicagram_associated.R
\name{gallicagram_associated}
\alias{gallicagram_associated}
\title{Word the most often associated with a n-gram in a Gallicagram corpus}
\usage{
gallicagram_associated(
  keyword,
  corpus = "lemonde",
  from = 1945,
  to = 2022,
  n_associated_words = 20,
  after = TRUE,
  stopwords = NULL
)
}
\arguments{
\item{keyword}{A character string. Keyword to search associations for.}

\item{corpus}{A character string. The corpus to search. Takes the following
values: "press" for historical press, "books" for Gallica books,
"lemonde" for Le Monde newspaper articles.}

\item{from}{A number. Starting year.}

\item{to}{A number. End year.}

\item{n_associated_words}{An integer. The number of most frequently
associated words to return.}

\item{after}{A boolean. Whether to consider only words following the keyword
and not those preceding. Set to \code{TRUE} by default.}

\item{stopwords}{A character vector of stopwords to remove.
Can for instance be \code{lsa::stopwords_fr}.
Removed after withdrawing apostrophes and letters preceding them.
If \code{NULL} does not remove any stopwords.}
}
\value{
A tibble. With the most frequent ngrams containing the
\code{keyword} mentioned (\code{ngram}), the \code{associated_word},
and the number of occurrences over the period (\code{n_occur}).
It also returns the input parameters
\code{keyword}, \code{corpus}, \code{from} and \code{to}.
}
\description{
Returns the word the most frequently following a given ngram over the period.
It can also include the words that precede by setting \code{after = FALSE}.
}
\details{
This function corresponds to the \code{Joker} route of the API,
accessed through the 'Joker' function on the Gallicagram app
and is analogous to the 'Joker' function on Ngram Viewer.

For instance "camarade" is often followed by "staline" or "khrouchtchev" in
Le Monde. The function returns the most frequent ngrams of the form
"camarade *". Setting \code{after = FALSE} also includes the most frequent
ngrams of the form "* camarade".

The keyword can be a 2-gram in the "books" and "press" corpora and a 3-gram
in the "lemonde" corpus. Searching the "press" corpus can require a long
running time.
}
\examples{
gallicagram_associated("camarade")
}
