% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gallicagram_associated.R
\name{gallicagram_associated}
\alias{gallicagram_associated}
\title{Words the most often associated with a keyword in a Gallicagram corpus}
\usage{
gallicagram_associated(
  keyword,
  corpus = "lemonde",
  from = 1945,
  to = 2022,
  n_results = 20,
  distance = "max",
  stopwords = rallicagram::stopwords_gallica[1:500]
)
}
\arguments{
\item{keyword}{A character string. Keyword to search. Can be up to a 3-gram
in the "books" and "press" corpora and a 4-gram in the "lemonde" corpus.}

\item{corpus}{A character string. The corpus to search. Takes the following
values: "press" for historical press, "books" for Gallica books,
"lemonde" for Le Monde newspaper articles.}

\item{from}{An integer. Starting year.}

\item{to}{An integer. End year.}

\item{n_results}{An integer. The number of most frequently
associated words to return. \code{n_results} can also be set to "all" to
return all the available results.}

\item{distance}{An integer or "max". The maximum distance,
in number of words, at which to look for words associated with the keyword.
If equal to "max", will be 3 for the Le Monde corpus and 2 otherwise.}

\item{stopwords}{A character vector of stopwords to remove.
The default is the vector of the 500 most frequent words in the Gallica
books dataset. We can change this number by passing
\code{stopwords_gallicca[1:300]} (for instance, for the 300 most frequent)
to the \code{stopwords} argument. Can also be \code{lsa::stopwords_fr}
If \code{NULL} does not remove any stopwords.}
}
\value{
A tibble. Containing the words the most frequently associated with the
\code{keyword} mentioned (\code{associated_word}),
and the number of occurrences over the period (\code{n_occur}).
It also returns the input parameters
\code{keyword}, \code{corpus}, \code{from} and \code{to}.
}
\description{
Returns the word the most frequently at a given number of words
(\code{distance}) from the keyword.
}
\details{
This functions calls the '\code{associated}' route of the API.

Note that the API route does not allow to search for associated words
after a punctuation mark.
For instance, in "son camarade, le chasseur, a", the function will
not count "chasseur" as a word associated with "camarade". The ngram
"son camarade le chasseur" is not in the database.
Thus, there might be less associated words at a longer distance
as it increases the probability of ngrams to be excluded from the database.

Apostrophes and letters preceding them are withdrawn from the dataset
(except for n' since they carry meaning)

Searching the "press" corpus can require a long running time.
}
\examples{
gallicagram_associated("camarade")
}
