% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gallicagram_associated_lexicon.R
\name{gallicagram_associated_lexicon}
\alias{gallicagram_associated_lexicon}
\title{Words the most often associated with a vector of keywords
in a Gallicagram corpus}
\usage{
gallicagram_associated_lexicon(
  lexicon,
  corpus = "lemonde",
  from = "earliest",
  to = "latest",
  n_results = 20,
  distance = "max",
  stopwords = rallicagram::stopwords_gallica[1:500]
)
}
\arguments{
\item{lexicon}{A character vector. Keywords to search.}

\item{corpus}{A character string. The corpus to search. The list of
available corpora can be found in the \code{list_corpora} dataset.}

\item{from}{An integer or "earliest". The starting year.
If set to "earliest", use the earliest date at which the data is reliable
for this corpus, as described in \code{list_corpora}.}

\item{to}{An integer or "latest". The end year.
If set to "latest", use the latest date at which the data is reliable
for this corpus, as described in \code{list_corpora}.}

\item{n_results}{An integer. The number of most frequently
associated words to return. \code{n_results} can also be set to "all" to
return all the available results.}

\item{distance}{An integer or "max". The maximum distance,
in number of words, at which to look for words associated with the keyword.
The max length for each corpus (distance + number of words in the keyword)
is described in the \code{max_length} column
of the \code{list_corpora} dataset.}

\item{stopwords}{A character vector of stopwords to remove.
The default is the vector of the 500 most frequent words in the Gallica
books dataset. We can change this number by passing
\code{stopwords_gallicca[1:300]} (for instance, for the 300 most frequent)
to the \code{stopwords} argument. Can also be \code{lsa::stopwords_fr}
If \code{NULL} does not remove any stopwords.}
}
\value{
A tibble. With the words the most frequently associated with any of
the keywords in the \code{lexicon} mentioned (\code{associated_word}),
the first \code{keyword} in this lexicon, typically the main one
and the number of co-occurrences between the keyword and the associated
word over the period (\code{n_co-occur}).
It also returns the input parameters
\code{keyword}, \code{corpus}, \code{from} and \code{to}.
}
\description{
Returns the word the most frequently at a given number of words
(\code{distance}) from any of the keywords in the vector.
}
\details{
This function sums the outputs of calls of \code{gallicagram_associated}
obtained for each keyword in the vector.

Can typically be used with the function \code{get_same_stem()}.

This function is only available for the three main corpora
(historical press, Gallica books, Le Monde newspaper).
}
\examples{
\dontrun{
  gallicagram_associated_lexicon(c("président", "présidentiel"))
}
}
