% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gallicagram_cooccur.R
\name{gallicagram_cooccur}
\alias{gallicagram_cooccur}
\title{Close co-occurrences of two keywords in a Gallicagram copus}
\usage{
gallicagram_cooccur(
  keywords,
  corpus = "press",
  from = 1789,
  to = 1950,
  resolution = "monthly"
)
}
\arguments{
\item{keywords}{A character vector of length 2
containing the two keyword to search.}

\item{corpus}{A character string. The corpus to search. Takes the following
values: "press" for historical press, "books" for Gallica books,
"lemonde" for Le Monde newspaper articles.}

\item{from}{A number. Starting year.}

\item{to}{A number. End year.}

\item{resolution}{A character string.
For Le Monde can be either "yearly", "monthly" or "daily".
For historical press can be either "yearly", "monthly".
For books can only only be. "yearly".}
}
\value{
A tibble. With the \code{keyword},
the number of occurrences (\code{n}), the total number of ngrams
over the period of a given observation (\code{total}),
the proportion of occurrences of the keyword over the period of
a given observation (\code{prop}),
the date at the beginning of the period of a given observation (\code{date}),
the \code{source}, the \code{resolution},
the \code{year} and
potentially the \code{month} and \code{day} of the observation.
}
\description{
Retrieves the number of close co-occurrences (less than 4 grams appart)
of two keywords in one of the corpora
(historical press, Gallica books, Le Monde newspaper) by year, month or day.
}
\examples{
gallicagram_cooccur(c("président", "menteur"))
}
