% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gallicagram_cooccur.R
\name{gallicagram_cooccur}
\alias{gallicagram_cooccur}
\title{Close co-occurrences of two keywords in a Gallicagram copus}
\usage{
gallicagram_cooccur(
  keyword_1,
  keyword_2,
  corpus = "lemonde",
  from = "earliest",
  to = "latest",
  resolution = "monthly",
  count_phrase = FALSE,
  cooccur_level = "grams"
)
}
\arguments{
\item{keyword_1}{A character string. One of the two keywords to search.}

\item{keyword_2}{A character string. The other keyword to search.}

\item{corpus}{A character string. The corpus to search. The list of
available corpora can be found in the \code{list_corpora} dataset.}

\item{from}{An integer or "earliest". The starting year.
If set to "earliest", use the earliest date at which the data is reliable
for this corpus, as described in \code{list_corpora}.}

\item{to}{An integer or "latest". The end year.
If set to "latest", use the latest date at which the data is reliable
for this corpus, as described in \code{list_corpora}.}

\item{resolution}{A character string.
For press and lemonde can be either "yearly" or "monthly".
For books can only be "yearly".}

\item{count_phrase}{If TRUE, counts the co-occurrences of each phrase
containing both keywords.
If FALSE, returns the number of times both keywords co-occur
in each resolution period.}

\item{cooccur_level}{character string. Either "grams" or "articles".
The level at which to look for co-occurences of the two keywords:
in 3-grams for "livres" and "presse"
and in 4-grams or articles for "lemonde".}
}
\value{
A tibble. With the \code{keyword},
the number of occurrences (\code{n_occur}) or co-occurrences
(\code{n_cooccur}), the total number of ngrams or aritcles
over the period (\code{n_total}),
the proportion of occurrences or co-occurrences of the keyword(s) over the
period of a given observation (\code{prop_occcur} or \code{prop_coocccur}),
either info about whether the total number is a number of grams or articles
(\code{n_of}) or about
the syntagma at which the co-occurrences are computed (\code{cooccur_level}),
the date at the beginning of the period of a given observation (\code{date}),
the \code{corpus}, the \code{resolution},
the \code{year} and
potentially the \code{month} and \code{day} of the observation.
}
\description{
Retrieves the proportion of close co-occurrences of two keywords in one of
the three main corpora (historical press, Gallica books, Le Monde newspaper)
by year or month.
}
\details{
Close co-occurrences correspond to the number of 3-grams
(4-grams in the Le Monde corpus) that contain the two keywords.

This function is only available for the three main corpora
(historical press, Gallica books, Le Monde newspaper).

It corresponds to the \code{Contain} route of the API.
}
\examples{
  gallicagram_cooccur("président", "mauvais")
}
