% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gallicagram_cooccur.R
\name{gallicagram_cooccur}
\alias{gallicagram_cooccur}
\title{Close co-occurrences of two keywords in a Gallicagram copus}
\usage{
gallicagram_cooccur(
  keywords,
  corpus = "lemonde",
  from = 1945,
  to = 2022,
  resolution = "monthly"
)
}
\arguments{
\item{keywords}{A character vector of length 2
containing the two keyword to search.}

\item{corpus}{A character string. The corpus to search. Takes the following
values: "press" for historical press, "books" for Gallica books,
"lemonde" for Le Monde newspaper articles.}

\item{from}{A number. Starting year.}

\item{to}{A number. End year.}

\item{resolution}{A character string.
For press and lemonde can be either "yearly" or "monthly".
For books can only be "yearly".}
}
\value{
A tibble. With the \code{keyword},
the number of occurrences (\code{n_occur}) or co-occurrences
(\code{n_cooccur}), the total number of ngrams over the period of a
given observation (\code{n_grams} or \code{n_ngrams}),
the proportion of occurrences or co-occurrences of the keyword(s) over the
period of a given observation (\code{prop_occcur} or \code{prop_coocccur}),
the date at the beginning of the period of a given observation (\code{date}),
the \code{source}, the \code{resolution},
the \code{year} and
potentially the \code{month} and \code{day} of the observation.
}
\description{
Retrieves the proportion of close co-occurrences of two keywords in one of
the corpora (historical press, Gallica books, Le Monde newspaper)
by year, month or day.
}
\details{
Close co-occurrences correspond to the number of 3-grams
(4-grams in the Le Monde corpus) that contain the two keywords.

This function corresponds to the \code{Contain} route of the API.
}
\examples{
gallicagram_cooccur(c("président", "mauvais"))
}
