% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gallicagram_cooccur_lexicon.R
\name{gallicagram_cooccur_lexicon}
\alias{gallicagram_cooccur_lexicon}
\title{Close co-occurrences of two lexicons in a Gallicagram copus}
\usage{
gallicagram_cooccur_lexicon(
  lexicon_1,
  lexicon_2,
  corpus = "lemonde",
  from = "earliest",
  to = "latest",
  resolution = "monthly"
)
}
\arguments{
\item{lexicon_1}{A character vector. One of the two lexicons to search.}

\item{lexicon_2}{A character vector. The other lexicon to search.}

\item{corpus}{A character string. The corpus to search. The list of
available corpora can be found in the \code{list_corpora} dataset.}

\item{from}{An integer or "earliest". The starting year.
If set to "earliest", use the earliest date at which the data is reliable
for this corpus, as described in \code{list_corpora}.}

\item{to}{An integer or "latest". The end year.
If set to "latest", use the latest date at which the data is reliable
for this corpus, as described in \code{list_corpora}.}

\item{resolution}{A character string.
For press and lemonde can be either "yearly" or "monthly".
For books can only be "yearly".}
}
\value{
A tibble. With the \code{keyword},
the number of occurrences (\code{n_occur}) or co-occurrences
(\code{n_cooccur}), the total number of ngrams over the period of a
given observation (\code{n_grams} or \code{n_ngrams}),
the proportion of occurrences or co-occurrences of the keyword(s) over the
period of a given observation (\code{prop_occcur} or \code{prop_coocccur}),
the date at the beginning of the period of a given observation (\code{date}),
the \code{corpus}, the \code{resolution},
the \code{year} and
potentially the \code{month} and \code{day} of the observation.
}
\description{
Retrieves the proportion of close co-occurrences of two lexicons in one of
the three main corpora (historical press, Gallica books, Le Monde newspaper)
by year or month.
}
\details{
Close co-occurrences correspond to the number of 3-grams
(4-grams in the Le Monde corpus) that contain each pair of keywords in the
two lexicons.

This function simply loops the function \code{gallicagram_cooccur} over each
word of each lexicon and sums the results. It can thus take some time to run.

This function is only available for the three main corpora
(historical press, Gallica books, Le Monde newspaper).
}
\examples{
gallicagram_cooccur_lexicon(c("président", "présidente"), c("mauvais", "nul"))
}
