% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gallicagram_with.R
\name{gallicagram_with}
\alias{gallicagram_with}
\title{ngrams that most frequently contain a keyword in a Gallicagram corpus over a
period}
\usage{
gallicagram_with(
  keyword,
  corpus = "lemonde",
  from = 1945,
  to = 2022,
  n_results = 20,
  after = FALSE,
  length = 2
)
}
\arguments{
\item{keyword}{A character string. Keyword to search.
The largest number of words the keyword can contain can be found
in the \code{resolution} column of the \code{list_corpora} dataset.}

\item{corpus}{A character string. The corpus to search. The list of
available corpora can be found in the \code{list_corpora} dataset.}

\item{from}{An integer. Starting year.}

\item{to}{An integer. End year.
The finest available resolution for the corpus selected can be found
in the \code{resolution} column of the \code{list_corpora} dataset.}

\item{n_results}{An integer. The number of most frequently
associated words to return. \code{n_results} can also be set to "all" to
return all the available results.}

\item{after}{A boolean. Whether to consider only words following the keyword
and not those preceding. Set to \code{FALSE} by default.}

\item{length}{An integer. The length of the ngrams considered.
Can be up to 3 in the "books" and "press" corpora and 4 in the
"lemonde" corpus.}
}
\value{
A tibble. With the \code{n_results} most frequent ngrams containing
the \code{keyword} searched (\code{ngram})
and the number of occurrences over the period (\code{n_occur}).
It also returns the input parameters
\code{keyword}, \code{corpus}, \code{from} and \code{to}.
}
\description{
Returns the most frequent ngrams containing a keyword over a given period.
}
\details{
This function corresponds to the \code{Joker} route of the API,
accessed through the 'Joker' function on the Gallicagram app.
When length = 1, it is analogous to the 'Joker' function on Ngram Viewer.

It is analogous to \code{gallicagram_with_month} but for a period instead of
a given month.

For instance "camarade" is often followed by "staline" or "khrouchtchev" in
Le Monde. The function returns the most frequent ngrams of the form
"camarade *" when setting \code{after = TRUE}. \code{after = FALSE} also
includes the most frequent ngrams of the form "* camarade".

Searching the "presse" corpus can require a long running time.
}
\examples{
\dontrun{
  gallicagram_with("camarade", length = 2)
}
}
