% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gallicagram_with_month.R
\name{gallicagram_with_month}
\alias{gallicagram_with_month}
\title{ngrams that most frequently contain a keyword in a Gallicagram corpus for a
specific month}
\usage{
gallicagram_with_month(
  keyword,
  corpus = "lemonde",
  year = 2022,
  month = 1,
  n_results = 20,
  after = FALSE,
  length = 2
)
}
\arguments{
\item{keyword}{A character string. Keyword to search. Can be up to a 3-gram
in the "books" and "press" corpora and a 4-gram in the "lemonde" corpus.}

\item{corpus}{A character string. The corpus to search. Takes the following
values: "press" for historical press, "books" for Gallica books,
"lemonde" for Le Monde newspaper articles.}

\item{year}{An integer. The year of interest.}

\item{month}{An integer. The month of the \code{year} of interest.}

\item{n_results}{An integer. The number of most frequently
associated words to return. \code{n_results} can also be set to "all" to
return all the available results.}

\item{after}{A boolean. Whether to consider only words following the keyword
and not those preceding. Set to \code{FALSE} by default.}

\item{length}{An integer. The length of the ngrams considered.
Can be up to 3 in the "books" and "press" corpora and 4 in the
"lemonde" corpus.}
}
\value{
A tibble. With the \code{n_results} most frequent ngrams containing
the \code{keyword} searched (\code{ngram})
and the number of occurrences over the period (\code{n_occur}).
It also returns the input parameters
\code{keyword}, \code{corpus}, \code{year} and \code{month}.
}
\description{
Returns the most frequent ngrams containing a keyword for a given month
}
\details{
This function corresponds to the \code{joker_month} route of the API.

It is analogous to \code{gallicagram_with} but for a precise month instead
of a given period.

For instance "camarade" is often followed by "staline" or "khrouchtchev" in
Le Monde. The function returns the most frequent ngrams of the form
"camarade *" when setting \code{after = TRUE}. \code{after = FALSE} also
includes the most frequent ngrams of the form "* camarade".

Searching the "press" corpus can require a long running time.
}
\examples{
gallicagram_with_month("camarade", length = 2)
}
