% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gallicagraph.R
\name{gallicagraph}
\alias{gallicagraph}
\title{Automatically produces the relevant graph}
\usage{
gallicagraph(data, color = NULL, labels_color = NA)
}
\arguments{
\item{data}{A data frame produced by one of the \code{gallicagram},
\code{gallicagram_coocur} or \code{gallicagram_associated}
functions or their lexicon counterparts
(or several of such data frames bound by rows).}

\item{color}{A variable to set colors for the graph}

\item{labels_color}{A character string. The color in which to print
the labels. Set to \code{NA} for no label.}
}
\value{
A graph describing the evolution of the proportion of
(co-)occurrences of one or several keywords in one or several corpora.
}
\description{
Takes as input data produced by one of the \code{gallicagram},
\code{gallicagram_coocur} or \code{gallicagram_associated}
functions or their lexicon counterparts and
produces the corresponding base graph.
}
\details{
This function can also be combined with faceting by adding for instance
\code{+ facet_wrap(~ keyword)} after calling the function.
}
\examples{
  gallicagram("président") |>
    gallicagraph()

  gallicagram_cooccur("président", "ancien") |>
    gallicagraph()

  gallicagram_associated("patate", from = 2020) |>
    gallicagraph()
}
