% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gallicagraph.R
\name{gallicagraph}
\alias{gallicagraph}
\title{Graph describing the evolution of the proportion of occurrences}
\usage{
gallicagraph(data, color = NULL)
}
\arguments{
\item{data}{A data frame produced by the \code{gallicagram} function
(or several of such data frames bound by rows).}

\item{color}{A variable to set colors for the graph}
}
\value{
A graph describing the evolution of the proportion of occurrences
of one or several keywords in one or several corpora.
}
\description{
Takes as input data produced by \code{gallicagram} and produces a graph
describing the evolution of the proportion of occurrences in this data set.
}
\details{
This function can also be combined with faceting by adding for instance
\code{+ facet_wrap(~ keyword)} after calling the function.
}
\examples{
gallicagram("président") |>
  gallicagraph()

gallicagram("président") |>
  rbind(gallicagram("république")) |>
  gallicagraph(color = keyword)

gallicagram("président") |>
  rbind(gallicagram("république")) |>
  gallicagraph() +
  facet_wrap(~ keyword)
}
