% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{list_corpora}
\alias{list_corpora}
\title{A data frame containing information about the available corpora}
\format{
A data frame with 17 rows and 7 variables:
\describe{
  \item{corpus}{Code name of the corpus.}
  \item{name_corpus}{Plain language name of the corpus.}
  \item{start_reliable}{The year at which the corpus starts being reliable.}
  \item{end_reliable}{The year at which the corpus stops being reliable.}
  \item{nb_words}{The number of words in the corpus.}
  \item{max_length}{The maximum length of ngrams available.}
  \item{resolution}{The finest available resolution (daily, monthly, yearly)}
}
}
\usage{
data("list_corpora")
}
\description{
It includes the code name of the corpus, its plain language name,
the years for which the data is reliable,
the number of words in each corpus,
the maximum length of the ngrams, and the resolution.
}
\examples{
\dontrun{
  # Load the dataset
  data("list_corpora")
}

}
\keyword{datasets}
