% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_param.R
\name{prepare_param}
\alias{prepare_param}
\title{Internal function to prepare the parameters to use them in the API call.}
\usage{
prepare_param(
  keyword,
  corpus,
  from,
  to,
  resolution = "yearly",
  n_of_level = "grams"
)
}
\arguments{
\item{keyword}{A character string specifying the keyword for analysis.}

\item{corpus}{A character string specifying the code of the corpus.}

\item{from}{An integer or "earliest" representing the starting year
for the analysis.}

\item{to}{An integer or "latest" representing the ending year
for the analysis.}

\item{resolution}{A character string specifying the desired resolution for
the analysis (e.g., "daily", "monthly", "yearly").}

\item{n_of_level}{A character string specifying the type of object to
compute the number of occurrences for or the level at which to compute
co-occurrences.}
}
\value{
A list.
Contains cleaned function parameters (`keyword`, `corpus`, `from`, `to`
and `resolution`).
}
\description{
Checks potential errors, translates the parameters, set parameters for
the API call.
}
