% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_param.R
\name{prepare_param}
\alias{prepare_param}
\title{Internal function to prepare the parameters to use them in the API call.}
\usage{
prepare_param(keyword, corpus, from, to, resolution)
}
\arguments{
\item{keyword}{A character string. Keyword to search.}

\item{corpus}{A character string. The corpus to search. Takes the following
values: "press" for historical press, "books" for Gallica books,
"lemonde" for Le Monde newspaper articles.}

\item{from}{A number. Starting year.}

\item{to}{A number. End year.}

\item{resolution}{A character string.
For Le Monde can be either "yearly", "monthly" or "daily".
For historical press can be either "yearly", "monthly".
For books can only only be. "yearly".}
}
\value{
A list.
Contains cleaned function parameters (`keyword`, `corpus` and `resolution`).
}
\description{
Internal function to prepare the parameters to use them in the API call.
}
