% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rallicagram_search.R
\name{rallicagram_search}
\alias{rallicagram_search}
\title{Retrieves the number of occurrences of a keyword in one of the corpora
(historical press, Gallica books, Le Monde newspaper)}
\usage{
rallicagram_search(
  keyword,
  corpus = "press",
  begin = 1789,
  end = 1950,
  frequency = "monthly"
)
}
\arguments{
\item{keyword}{A character string. Keyword to search.}

\item{corpus}{A character string. The corpus to search. Takes the following
values: "press" for historical press, "books" for Gallica books,
"lemonde" for Le Monde newspaper articles.}

\item{begin}{A number. Starting year.}

\item{end}{A number. End year.}

\item{frequency}{A number. End year.}
}
\value{
A tibble.
}
\description{
Retrieves the number of occurrences of a keyword in one of the corpora
(historical press, Gallica books, Le Monde newspaper)
}
\examples{
rallicagram_search("président")
}
