% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_gallicagram.R
\name{tidy_gallicagram}
\alias{tidy_gallicagram}
\title{Internal function to tidy the output of the API call.}
\usage{
tidy_gallicagram(url, corpus, resolution_en)
}
\arguments{
\item{url}{A url to pass the Gallicagram API}

\item{corpus}{A character string. The corpus to search. The list of
available corpora can be found in the \code{list_corpora} dataset.}

\item{resolution_en}{A character string. Can only be "daily", "monthly" or
"yearly".}
}
\value{
A tibble. With the \code{keyword},
the number of occurrences (\code{n_occur}) or co-occurrences
(\code{n_cooccur}), the total number of ngrams over the period of a
given observation (\code{n_grams} or \code{n_ngrams}),
the proportion of occurrences or co-occurrences of the keyword(s) over the
period of a given observation (\code{prop_occcur} or \code{prop_coocccur}),
the date at the beginning of the period of a given observation (\code{date}),
the \code{corpus}, the \code{resolution},
the \code{year} and
potentially the \code{month} and \code{day} of the observation.
}
\description{
Turns the Gallicagram API data into a tibble, renames the variables and
add the input parameters to the output tibble.
}
